requirements.txt
flask
discordbot.py
gunicorn
python-dotenv
petpetgif